Added IdNumber idetentefier to report
I've added the idnumber to the reports tab/csv file which is currently optional in moodle, but which a lot of universities use as a specific number to identify a student.

I've also split fullname (again, only in reports) to name and surname so people can sort people easier (usually by surname, but now they have the ability to also do it by firstname).